Modify speech-to-text
